requirements.txt
Flask
google-generativeai
python-docx
pyyaml
Flask-Session
beautifulsoup4
markdown